refactor(storage): move iceberg report helpers out of compactor mod
